Win32: Fixed compilation
git-svn-id: file:///home/michael/svn/yuni/trunk@1090 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/file/stream.cpp
@@ -67,9 +67,9 @@ namespace File
 	{
 		switch (origin)
 		{
-			case seekOriginBegin:   return (0 == ::fseek(pFd, offset, SEEK_SET));
-			case seekOriginCurrent: return (0 == ::fseek(pFd, offset, SEEK_CUR));
-			case seekOriginEnd:     return (0 == ::fseek(pFd, offset, SEEK_END));
+			case seekOriginBegin:   return (0 == ::fseek(pFd, (long)offset, SEEK_SET));
+			case seekOriginCurrent: return (0 == ::fseek(pFd, (long)offset, SEEK_CUR));
+			case seekOriginEnd:     return (0 == ::fseek(pFd, (long)offset, SEEK_END));
 		}
 		return false;
 	}
